endpoint /lootboxes working

+++ src/main/java/com/cs2drop/repository/LootboxRepository.java
@@ -3,4 +3,4 @@ package com.cs2drop.repository;
 import com.cs2drop.entity.Lootbox;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LootboxRepository extends JpaRepository<Lootbox, Long> {}
+public interface LootboxRepository extends JpaRepository<Lootbox, Integer> {}
